# dist/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: WikiFactMine-api
# during dev, should point to your local machine
host: tools.wmflabs.org
# basePath prefixes all resource paths
basePath: /wikifactmine-api/api
#
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
  /date/{date}:
    # binds a127 app logic to a route
    x-swagger-router-controller: wikifactmine-api
    get:
      description: Returns Facts for a given date
      # used as the method name of the controller
      operationId: date
      parameters:
        - name: date
          in: path
          description: The ingestion load date facts should be returned from
          required: true
          type: string
          format: date
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/FactsResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /item/{item}:
    x-swagger-router-controller: wikifactmine-api
    get:
      description: Returns Facts for a given Wikidata Item
      # used as the method name of the controller
      operationId: item
      parameters:
        - name: item
          in: path
          description: The Wikidata Item facts should be related to
          required: true
          type: string
          pattern: "^Q[0-9]*$"
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/FactsResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /2item/{item1}/{item2}:
    x-swagger-router-controller: wikifactmine-api
    get:
      description: Returns Facts from Papers where the two specified items cooccur
      # used as the method name of the controller
      operationId: item
      parameters:
        - name: item1
          in: path
          description: The First Wikidata Item facts should be related to
          required: true
          type: string
          pattern: "^Q[0-9]*$"
        - name: item2
          in: path
          description: The Second Wikidata Item facts should be related to
          required: true
          type: string
          pattern: "^Q[0-9]*$"
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/FactsResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /swagger:
    x-swagger-pipe: swagger_raw
# complex objects have schema definitions
definitions:
  FactsResponse:
    required:
      - message
    properties:
      message:
        type: string
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string